categorizer_backend/requirements.txt
Django>=3.0,<4.0
ipython
psycopg2>=2.8
django-rest-framework
django-cors-headers
